web_auto/requirements.txt
selenium
aiohttp
beautifulsoup4
google-generativeai
python-dotenv
requests
